Reading Transactions from file!

+++ zenon-fraud-detector/src/main/java/br/com/zenon/fraud/Transaction.java
@@ -2,7 +2,7 @@ package br.com.zenon.fraud;
 
 import java.math.BigDecimal;
 
-public record Transaction(Long step, TransactionTypes type, BigDecimal amount, Customer origin,
+public record Transaction(int step, TransactionTypes type, BigDecimal amount, Customer origin,
                           Customer destination,
                           boolean isFraud,
                           boolean isFlaggedFraud) {
